Now Post is working fine
Changed the mappting annotations. Have changed from RequestMapping to PostMapping and GetMapping. The Applications is working fine. In the services also now I am using a Java util class to hold on the members. When the parameter validation gets added might be use full to use a map. Still have not tested what happens when the JSON is mal formed