Now able to add error handling in the API
As of now doing this by raising an Excpetion ResponseStatusException and passing in the HTTP status code and error message to go with it. The reponse for error code looks fine but should remove the call stack in case of production ready code which is still shown in this implementation.